Party container added
So.. problem.
A lot to build is kind of dictated on the backend and pulling. I can hardcode it but not efficient. We should focus on specific area.

Also couldn't get icons to work. I feel pattern recognition is confusing for users (it was for me). we'll bring it up at meeting